+++ app/helpers/recording_studio_terms_and_conditions/continue_notice_helper.rb
@@ -4,6 +4,7 @@ module RecordingStudioTermsAndConditions
   module ContinueNoticeHelper
     include TermsContextHelper
     include AgreeCopyHelper
+    include DocumentButtonHelper
 
     CONTINUE_NOTICE_SIZES = { xs: "text-xs", sm: "text-sm" }.freeze